move travelthru logic to a separate module. todo: split off the report writing from the collection of units.
