+++ src/report.h
@@ -1,3 +1,4 @@
+#pragma once
 /* 
  +-------------------+  Christian Schlittchen <corwin@amber.kn-bremen.de>
  |                   |  Enno Rehling <enno@eressea.de>
@@ -11,6 +12,9 @@
  */
 #ifndef H_GC_REPORT
 #define H_GC_REPORT
+
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -21,7 +25,8 @@ extern "C" {
     void register_nr(void);
     void report_cleanup(void);
     void write_spaces(struct stream *out, size_t num);
-    void write_travelthru(struct stream *out, const struct region * r, const struct faction * f);
+
+    void paragraph(struct stream *out, const char *str, ptrdiff_t indent, int hanging_indent, char marker);
 #ifdef __cplusplus
 }
 #endif
